fixed planningoverview
