--- src/screens/planningOverzichtDev/Styles.js
@@ -1,41 +1,0 @@
-import { StyleSheet, Platform, StatusBar } from "react-native";
-
-export default StyleSheet.create({
-	SafeAreaView: {
-		paddingTop: Platform.OS === "android" ? StatusBar.currentHeight : 0,
-		backgroundColor: "#009BAA",
-		flex: 1,
-	},
-    row: {
-		flexDirection: 'row', 
-        paddingVertical: 5,
-		backgroundColor: "white",
-		borderWidth: 2,
-        borderRadius: 9,
-        justifyContent: 'center',
-		textAlign: 'left',
-		fontWeight: 'bold',
-		justifyContent: 'space-between',
-		marginBottom: 10,
-		
-
-	},
-	sectionHeader: {
-		backgroundColor: "#009BAA",
-		color: 'white',
-		
-	},
-	sectionList: {
-		flex: 1,
-		paddingHorizontal: 20,
-        paddingVertical: 10,
-		textAlign: 'left',
-      	fontSize: 20,
-		backgroundColor: "#009BAA",
-		
-		
-	},
-	icon: {
-		textAlign: 'right',
-	},
-});
